updated to support doubles for levels, in the case of a thermostat

+++ zVirtualClient/Models/Device.cs
@@ -10,7 +10,7 @@ namespace zVirtualClient.Models
         public int id { get; set; }
         public string name { get; set; }
         public string on_off { get; set; }
-        public int level { get; set; }
+        public double level { get; set; }
         public string level_txt { get; set; }
         public string type { get; set; }
         public string plugin_name { get; set; }
